+++ swift/Sources/SignalFfi/signal_ffi.h
@@ -3147,6 +3147,12 @@ SignalFfiError* signal_authenticated_chat_connection_set_device_name(
   uint8_t device_id,
   SignalBorrowedBuffer encrypted_name
 );
+SignalFfiError* signal_authenticated_chat_connection_set_discoverable_by_phone_number(
+  SignalCPromisebool* promise,
+  SignalConstPointerTokioAsyncContext async_runtime,
+  SignalConstPointerAuthenticatedChatConnection chat,
+  bool discoverable
+);
 SignalFfiError* signal_authenticated_chat_connection_set_push_token_apns(
   SignalCPromisebool* promise,
   SignalConstPointerTokioAsyncContext async_runtime,