+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 v0.97.4
 
 - AuthAccountsService.setRegistrationLock() typed API
+- AuthAccountsService.setDiscoverableByPhoneNumber() typed API
 - UnauthBackupsService.copyBackupMedia() typed API
 - node: Add full-er service SVR2 APIs that performs necessary crypto operations inside libsignal.